@@ -1,10 +1,11 @@
 {
   "name": "vue3-lib-starter",
-  "version": "0.0.1",
-  "description": "Starter pack for create library on vue 3, vite, typescript, sass",
+  "version": "0.0.2",
+  "private": true,
+  "description": "Starter pack for create library on vue 3, vite, typescript, sass and vitest",
   "main": "./dist/my-lib.umd.js",
   "module": "./dist/my-lib.es.js",
-  "types": "./dist/index.d.ts",
+  "types": "./dist/src/index.d.ts",
   "exports": {
     ".": {
       "import": "./dist/my-lib.es.js",
@@ -16,38 +17,42 @@
   ],
   "scripts": {
     "dev": "vite",
-    "dev:host": "vite --host",
-    "build": "rimraf dist && run-p type-check build-only",
+    "build": "run-p type-check \"build-only {@}\" --",
+    "preview": "vite preview",
+    "test:unit": "vitest",
     "build-only": "vite build",
-    "type-check": "vue-tsc --noEmit --composite false",
+    "type-check": "vue-tsc --noEmit -p tsconfig.vitest.json --composite false",
     "lint": "eslint . --ext .vue,.ts",
     "lint:fix": "eslint . --ext .vue,.ts --fix",
     "format": "prettier -c .",
     "format:fix": "prettier -w .",
-    "check": "npm run lint && npm run format"
+    "check": "npm run type-check && npm run lint && npm run format"
   },
-  "peerDependencies": {
-    "vue": "^3.2.25"
+  "dependencies": {
+    "vue": "^3.3.4"
   },
   "devDependencies": {
-    "@rushstack/eslint-patch": "^1.2.0",
-    "@types/node": "^18.15.3",
-    "@vitejs/plugin-vue": "^4.0.0",
-    "@vue/eslint-config-prettier": "^7.1.0",
-    "@vue/eslint-config-typescript": "^11.0.2",
-    "@vue/tsconfig": "^0.1.3",
-    "autoprefixer": "^10.4.14",
-    "eslint": "^8.36.0",
-    "eslint-plugin-vue": "^9.9.0",
-    "npm-run-all": "^4.1.5",
-    "prettier": "^2.8.4",
-    "rimraf": "^4.4.0",
-    "sass": "^1.59.3",
-    "typescript": "^4.9.5",
-    "vite": "^4.1.4",
-    "vite-plugin-dts": "^2.1.0",
-    "vue": "^3.2.47",
-    "vue-router": "^4.1.6",
-    "vue-tsc": "^1.2.0"
+    "@rushstack/eslint-patch": "^1.3.3",
+    "@tsconfig/node18": "^18.2.2",
+    "@types/jsdom": "^21.1.3",
+    "@types/node": "^20.8.9",
+    "@vitejs/plugin-vue": "^4.4.0",
+    "@vue/eslint-config-prettier": "^8.0.0",
+    "@vue/eslint-config-typescript": "^12.0.0",
+    "@vue/test-utils": "^2.4.1",
+    "@vue/tsconfig": "^0.4.0",
+    "autoprefixer": "^10.4.16",
+    "eslint": "^8.49.0",
+    "eslint-plugin-vue": "^9.17.0",
+    "jsdom": "^22.1.0",
+    "npm-run-all2": "^6.1.1",
+    "prettier": "^3.0.3",
+    "sass": "^1.69.5",
+    "typescript": "~5.2.0",
+    "vite": "^4.4.11",
+    "vite-plugin-dts": "^3.6.2",
+    "vitest": "^0.34.6",
+    "vue-router": "^4.2.5",
+    "vue-tsc": "^1.8.19"
   }
 }
