@@ -7,12 +7,11 @@ module.exports = {
     'plugin:vue/vue3-essential',
     'eslint:recommended',
     '@vue/eslint-config-typescript',
-    '@vue/eslint-config-prettier',
+    '@vue/eslint-config-prettier/skip-formatting',
   ],
   parserOptions: {
     ecmaVersion: 'latest',
   },
-  ignorePatterns: ['**/*.json'],
   rules: {
     semi: ['error', 'always'],
     quotes: ['error', 'single'],
@@ -43,6 +42,5 @@ module.exports = {
         ignore: [],
       },
     ],
-    'no-console': 'warn',
   },
 };
